add: unit tests for `Env` struct
