+++ src/recipe/jinja.rs
@@ -2,6 +2,7 @@
 //!
 use std::{collections::BTreeMap, str::FromStr};
 
+use minijinja::value::Object;
 use minijinja::{Environment, Value};
 use rattler_conda_types::Version;
 
@@ -217,6 +218,119 @@ fn set_jinja(config: &SelectorConfig) -> minijinja::Environment<'static> {
     env
 }
 
+#[derive(Debug)]
+pub(crate) struct Env;
+impl std::fmt::Display for Env {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str("Env")
+    }
+}
+
+impl Object for Env {
+    fn kind(&self) -> minijinja::value::ObjectKind<'_> {
+        minijinja::value::ObjectKind::Plain
+    }
+
+    fn call_method(
+        &self,
+        _state: &minijinja::State,
+        name: &str,
+        args: &[Value],
+    ) -> Result<Value, minijinja::Error> {
+        match name {
+            "get" => {
+                let mut args = args.iter();
+                let Some(arg) = args.next() else {
+                    return Err(minijinja::Error::new(
+                        minijinja::ErrorKind::MissingArgument,
+                        "`get` requires at least one argument",
+                    ));
+                };
+                if args.next().is_some() {
+                    return Err(minijinja::Error::new(
+                        minijinja::ErrorKind::InvalidOperation,
+                        "`get` only accepts one argument",
+                    ));
+                }
+                let Some(key) = arg.as_str() else {
+                    return Err(minijinja::Error::new(
+                        minijinja::ErrorKind::InvalidOperation,
+                        "`get` requires a string argument",
+                    ));
+                };
+                match std::env::var(key) {
+                    Ok(r) => Ok(Value::from(r)),
+                    Err(e) => Err(minijinja::Error::new(
+                        minijinja::ErrorKind::InvalidOperation,
+                        e.to_string(),
+                    )),
+                }
+            }
+            "get_default" => {
+                let mut args = args.iter();
+                let Some(arg) = args.next() else {
+                    return Err(minijinja::Error::new(
+                        minijinja::ErrorKind::MissingArgument,
+                        "`get_default` requires at least two arguments",
+                    ));
+                };
+                let Some(key) = arg.as_str() else {
+                    return Err(minijinja::Error::new(
+                        minijinja::ErrorKind::InvalidOperation,
+                        "`get_default` requires string arguments",
+                    ));
+                };
+                let Some(arg) = args.next() else {
+                    return Err(minijinja::Error::new(
+                        minijinja::ErrorKind::MissingArgument,
+                        "`get_default` requires at least two arguments",
+                    ));
+                };
+                if args.next().is_some() {
+                    return Err(minijinja::Error::new(
+                        minijinja::ErrorKind::InvalidOperation,
+                        "`get_default` only accepts two arguments",
+                    ));
+                }
+                let Some(default) = arg.as_str() else {
+                    return Err(minijinja::Error::new(
+                        minijinja::ErrorKind::InvalidOperation,
+                        "`get_default` requires string arguments",
+                    ));
+                };
+                let ret = std::env::var(key).unwrap_or_else(|_| default.to_string());
+                Ok(Value::from(ret))
+            }
+            "exists" => {
+                let mut args = args.iter();
+                let Some(arg) = args.next() else {
+                    return Err(minijinja::Error::new(
+                        minijinja::ErrorKind::MissingArgument,
+                        "`exists` requires at least one argument",
+                    ));
+                };
+                if args.next().is_some() {
+                    return Err(minijinja::Error::new(
+                        minijinja::ErrorKind::InvalidOperation,
+                        "`exists` only accepts one argument",
+                    ));
+                }
+                let Some(key) = arg.as_str() else {
+                    return Err(minijinja::Error::new(
+                        minijinja::ErrorKind::InvalidOperation,
+                        "`exists` requires a string argument",
+                    ));
+                };
+                Ok(Value::from(std::env::var(key).is_ok()))
+            }
+            name => Err(minijinja::Error::new(
+                minijinja::ErrorKind::UnknownMethod,
+                format!("object has no method named {name}"),
+            )),
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use rattler_conda_types::Platform;
@@ -399,5 +513,61 @@ mod tests {
         assert!(!jinja.eval("cmp(python, '!=3.7')").expect("test 4").is_true());
         assert!(!jinja.eval("cmp(python, '<3.7')").expect("test 5").is_true());
         assert!(!jinja.eval("cmp(python, '>3.5,<3.7')").expect("test 6").is_true());
+    }
+
+    fn with_env((key, value): (impl AsRef<str>, impl AsRef<str>), f: impl Fn() -> ()) {
+        if let Ok(old_value) = std::env::var(key.as_ref()) {
+            std::env::set_var(key.as_ref(), value.as_ref());
+            f();
+            std::env::set_var(key.as_ref(), old_value);
+        } else {
+            std::env::set_var(key.as_ref(), value.as_ref());
+            f();
+            std::env::remove_var(key.as_ref());
+        }
+    }
+
+    #[test]
+    fn eval_env() {
+        let options = SelectorConfig {
+            target_platform: Platform::Linux64,
+            build_platform: Platform::Linux64,
+            variant: Default::default(),
+            hash: None,
+        };
+        let jinja = Jinja::new(options);
+
+        with_env(("RANDOM_JINJA_ENV_VAR", "false"), || {
+            assert_eq!(
+                jinja
+                    .eval("env.get('RANDOM_JINJA_ENV_VAR')")
+                    .expect("test 1")
+                    .as_str(),
+                Some("false")
+            );
+            assert!(jinja.eval("env.get('RANDOM_JINJA_ENV_VAR2')").is_err());
+            assert_eq!(
+                jinja
+                    .eval("env.get_default('RANDOM_JINJA_ENV_VAR', 'true')")
+                    .expect("test 3")
+                    .as_str(),
+                Some("false")
+            );
+            assert_eq!(
+                jinja
+                    .eval("env.get_default('RANDOM_JINJA_ENV_VAR2', 'true')")
+                    .expect("test 4")
+                    .as_str(),
+                Some("true")
+            );
+            assert!(jinja
+                .eval("env.exists('RANDOM_JINJA_ENV_VAR')")
+                .expect("test 5")
+                .is_true());
+            assert!(!jinja
+                .eval("env.exists('RANDOM_JINJA_ENV_VAR2')")
+                .expect("test 6")
+                .is_true());
+        });
     }
 }
