+++ src/selectors.rs
@@ -1,124 +1,10 @@
 use std::{collections::BTreeMap, str::FromStr};
 
-use minijinja::{
-    value::{Object, Value},
-    Environment,
-};
+use crate::recipe::jinja::Env;
+
+use minijinja::{value::Value, Environment};
 use rattler_conda_types::{Platform, Version, VersionSpec};
 use serde_yaml::Value as YamlValue;
-
-#[derive(Debug)]
-struct Env;
-impl std::fmt::Display for Env {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str("Env")
-    }
-}
-
-impl Object for Env {
-    fn kind(&self) -> minijinja::value::ObjectKind<'_> {
-        minijinja::value::ObjectKind::Plain
-    }
-
-    fn call_method(
-        &self,
-        _state: &minijinja::State,
-        name: &str,
-        args: &[Value],
-    ) -> Result<Value, minijinja::Error> {
-        match name {
-            "get" => {
-                let mut args = args.iter();
-                let Some(arg) = args.next() else {
-                    return Err(minijinja::Error::new(
-                        minijinja::ErrorKind::MissingArgument,
-                        "`get` requires at least one argument",
-                    ));
-                };
-                if args.next().is_some() {
-                    return Err(minijinja::Error::new(
-                        minijinja::ErrorKind::InvalidOperation,
-                        "`get` only accepts one argument",
-                    ));
-                }
-                let Some(key) = arg.as_str() else {
-                    return Err(minijinja::Error::new(
-                        minijinja::ErrorKind::InvalidOperation,
-                        "`get` requires a string argument",
-                    ));
-                };
-                match std::env::var(key) {
-                    Ok(r) => Ok(Value::from(r)),
-                    Err(e) => Err(minijinja::Error::new(
-                        minijinja::ErrorKind::InvalidOperation,
-                        e.to_string(),
-                    )),
-                }
-            }
-            "get_default" => {
-                let mut args = args.iter();
-                let Some(arg) = args.next() else {
-                    return Err(minijinja::Error::new(
-                        minijinja::ErrorKind::MissingArgument,
-                        "`get_default` requires at least two arguments",
-                    ));
-                };
-                let Some(key) = arg.as_str() else {
-                    return Err(minijinja::Error::new(
-                        minijinja::ErrorKind::InvalidOperation,
-                        "`get_default` requires string arguments",
-                    ));
-                };
-                let Some(arg) = args.next() else {
-                    return Err(minijinja::Error::new(
-                        minijinja::ErrorKind::MissingArgument,
-                        "`get_default` requires at least two arguments",
-                    ));
-                };
-                if args.next().is_some() {
-                    return Err(minijinja::Error::new(
-                        minijinja::ErrorKind::InvalidOperation,
-                        "`get_default` only accepts two arguments",
-                    ));
-                }
-                let Some(default) = arg.as_str() else {
-                    return Err(minijinja::Error::new(
-                        minijinja::ErrorKind::InvalidOperation,
-                        "`get_default` requires string arguments",
-                    ));
-                };
-                let ret = std::env::var(key).unwrap_or_else(|_| default.to_string());
-                Ok(Value::from(ret))
-            }
-            "exists" => {
-                let mut args = args.iter();
-                let Some(arg) = args.next() else {
-                    return Err(minijinja::Error::new(
-                        minijinja::ErrorKind::MissingArgument,
-                        "`exists` requires at least one argument",
-                    ));
-                };
-                if args.next().is_some() {
-                    return Err(minijinja::Error::new(
-                        minijinja::ErrorKind::InvalidOperation,
-                        "`exists` only accepts one argument",
-                    ));
-                }
-                let Some(key) = arg.as_str() else {
-                    return Err(minijinja::Error::new(
-                        minijinja::ErrorKind::InvalidOperation,
-                        "`exists` requires a string argument",
-                    ));
-                };
-                Ok(Value::from(std::env::var(key).is_ok()))
-            }
-            name => Err(minijinja::Error::new(
-                minijinja::ErrorKind::UnknownMethod,
-                format!("object has no method named {name}"),
-            )),
-        }
-    }
-}
 
 #[derive(Clone, Debug)]
 pub struct SelectorConfig {
